+++ CHANGELOG.md
@@ -17,6 +17,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
   [#698](https://github.com/MyIntervals/emogrifier/pull/698))
 
 ### Changed
+- Move utility classes to the `Pelago\Emogrifier\Utilities` namespace
+  ([#712](https://github.com/MyIntervals/emogrifier/pull/712))
 - Make the `$css` parameter of the `inlineCss` method optional
   ([#700](https://github.com/MyIntervals/emogrifier/pull/700))
 - Update the development dependencies